requirements.txt
# PyTorch with CUDA 11.8
torch==2.1.2
torchvision==0.16.2
--extra-index-url https://download.pytorch.org/whl/cu118

# Other Python packages
edt==2.4.1
elf==0.5.2
h5py==3.11.0
itk==5.3.0
lightning==2.1.3
monai==1.3.2
mrcfile==1.5.3
numpy==1.24.4
pandas==2.0.3
scipy==1.9.1
plotly==5.24.1
PyYAML==6.0.2
scikit_learn==1.3.2
scipy==1.9.1
scikit-image==0.20.0
edt==2.4.1
umap-learn==0.5.7